+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.2.8
 
+* Added [Azure Key Vault support](https://github.com/databrickslabs/terraform-provider-databricks/pull/381) for databricks_secret_scope for Azure CLI authenticated users
 * Added support for pinning clusters (issue #113)
 * Internal: API for retrieval of the cluster events
 * Fixed [single node clusters](https://docs.databricks.com/clusters/single-node.html) support by allowing [`num_workers` to be `0`](https://github.com/databrickslabs/terraform-provider-databricks/pull/375)